Added embedding code
